Module libfim::config::options

A module for handling internal configuration options.

There are three different types of options: boolean, numeric, and string. These can be set by the user through configuration files or in-fim commands (eventually, not right now).

Structs

OptionFactory

Struct that creates an Options object.

Options

Struct that represent the collection of internal configuration options.

Enums

LayoutType

Enum that represents different keyboard layouts.

LineNumbers

Enum that represents the different types of line numbers fim can use.

OptionParseError

Enum for containing errors that might occur in parsing option lines.