Trait libfim::context::Context

pub trait Context {
    fn setup(&mut self, _ed: &mut Editor<'_>) -> Result<bool> { ... }

    fn forward(
        &mut self, 
        _ed: &mut Editor<'_>, 
        _event: KeyEvent
    ) -> Result<Option<ContextMessage>> { ... }

    fn receive(
        &mut self, 
        _ed: &mut Editor<'_>, 
        _arg: ContextMessage
    ) -> Result<Option<ContextMessage>> { ... }
}

Trait to represent contexts.

See module-level documentation for more information.

Provided methods

fn setup(&mut self, _ed: &mut Editor<'_>) -> Result<bool>

Function to setup the Context when it becomes the active context.

Can cause the Context to ‘return’ and be popped off the context stack, if this function returns Ok(true). If this function returns Ok(false), the Context has not ‘returned’.

fn forward(
&mut self,
_ed: &mut Editor<'_>,
_event: KeyEvent
) -> Result<Option<ContextMessage>>

Accepts forwarded key presses.

Can cause the Context to ‘return’, if this function returns Ok(Some(c)), where c is a ContextMessage.

fn receive(
&mut self,
_ed: &mut Editor<'_>,
_arg: ContextMessage
) -> Result<Option<ContextMessage>>

Receives the return value of the Context above it on the context stack.

Can cause the Context to ‘return’, if this function returns Ok(Some(c)) where c is a ContextMessage.