Trait libfim::layout::Layout

pub trait Layout {
    fn from_qwerty(&self, qwerty_press: u8) -> u8;

    fn to_qwerty(&self, layout_press: u8) -> u8;

    fn from_qwerty_keycode(&self, qwerty_press: KeyCode) -> KeyCode { ... }

    fn to_qwerty_keycode(&self, layout_press: KeyCode) -> KeyCode { ... }
}

An interface for keyboard layouts.

Required methods

fn from_qwerty(&self, qwerty_press: u8) -> u8

Translate a QWERTY key press into a key press from this layout, by keyboard position. For example, a QWERTY ‘s’ translates into a Dvorak ‘o’ because they are located at the same place on the keyboard.

fn to_qwerty(&self, layout_press: u8) -> u8

Translate a key press from this layout into a QWERTY key press, by keyboard position. For example, a Dvorak ‘e’ translates into a QWERTY ‘d’ because they are located at the same place on the keyboard.

Provided methods

fn from_qwerty_keycode(&self, qwerty_press: KeyCode) -> KeyCode

Translate a QWERTY KeyCode into a KeyCode from this layout, by keyboard position. The default behavior is to return Self::from_qwerty() wrapped in a KeyCode on the enclosed character if it is an ASCII character, and otherwise return the argument.

fn to_qwerty_keycode(&self, layout_press: KeyCode) -> KeyCode

Translate a KeyCode from this layout into a QWERTY KeyCode, by keyboard position. The default behavior is to return Self::to_qwerty() wrapped in a KeyCode on the enclosed character if it is an ASCII character, and otherwise return the argument.